[CONNECTOR] rename FtpSource to Importer
