--- connector/src/main/java/org/astraea/connector/ftp/FtpSinkConnector.java
@@ -1,59 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.astraea.connector.ftp;
-
-import java.util.ArrayList;
-import java.util.List;
-import org.astraea.common.Configuration;
-import org.astraea.connector.Definition;
-import org.astraea.connector.SinkConnector;
-import org.astraea.connector.SinkTask;
-
-public class FtpSinkConnector extends SinkConnector {
-
-  private Configuration cons;
-
-  @Override
-  protected void init(Configuration configuration) {
-    this.cons = configuration;
-  }
-
-  @Override
-  protected Class<? extends SinkTask> task() {
-    return FtpSinkTask.class;
-  }
-
-  @Override
-  protected List<Configuration> takeConfiguration(int maxTasks) {
-    List<Configuration> configs = new ArrayList<>();
-    for (int i = 0; i < maxTasks; i++) {
-      configs.add(cons);
-    }
-    return configs;
-  }
-
-  @Override
-  protected List<Definition> definitions() {
-    return List.of(
-        Definition.builder()
-            .name("ftp")
-            .type(Definition.Type.STRING)
-            .defaultValue(null)
-            .documentation("test")
-            .build());
-  }
-}
